fix(migrate): align FK column types in 119, 120, 150 with target tables
Audit of every FOREIGN KEY in migrations 097-153 against the actual
column types in the deployed schema (phparm.sql) surfaced three more
type mismatches that would trip errno 150 on a fresh apply:

  119_pm_generations: schedule_id, plan_id, ticket_id were declared
    BIGINT UNSIGNED but pm_schedules.id, pm_plans.id, tickets.id are
    INT UNSIGNED. pm_generations.id stays BIGINT UNSIGNED — it's an
    append-only audit log and nothing FKs back to it.

  120_pm_generations_consumption: consumption_entitlement_id and
    consumption_ledger_id were BIGINT UNSIGNED while their referent
    tables (contract_entitlements, contract_consumption_ledger) are
    INT UNSIGNED. No explicit FK is declared in this file, but the
    columns are intended to FK and were already inconsistent.

  150_reporting_and_scheduled_reports: scheduled_reports.created_by
    was plain signed INT — the FK to users(id) failed with errno 150.
    The earlier sed fix only caught columns ending in _id; created_by
    slipped through.

+++ database/migrations/119_pm_generations.sql
@@ -11,9 +11,9 @@
 
 CREATE TABLE IF NOT EXISTS pm_generations (
     id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
-    schedule_id BIGINT UNSIGNED NOT NULL,
-    plan_id BIGINT UNSIGNED NOT NULL,
-    ticket_id BIGINT UNSIGNED NULL,
+    schedule_id INT UNSIGNED NOT NULL,
+    plan_id INT UNSIGNED NOT NULL,
+    ticket_id INT UNSIGNED NULL,
     due_at DATE NOT NULL,
     generated_at DATETIME NOT NULL,
     status VARCHAR(20) NOT NULL DEFAULT 'generated',
